+++ src/MCAL/GPIO/STM32F401cc_MCAL_GPIO.c
@@ -191,7 +191,7 @@ Sys_enuErrorStates_t GPIO_Set_PinValue(void *PtrToGPIOPort,u32 GPIO_PINnum, u32 
  * @return Sys_enuErrorStates_t
  * @warning GPIO Pin must be initialized first using GPIO_Init_Pin()  
  */
-Sys_enuErrorStates_t GPIO_Get_PinValue(void *PtrToGPIOPort,u32 GPIO_PINnum, u8 *PtrToPinstatus)
+Sys_enuErrorStates_t GPIO_Get_PinValue(void *PtrToGPIOPort,u32 GPIO_PINnum, u32 *PtrToPinstatus)
 {
     Sys_enuErrorStates_t Error_Status=NOT_OK;
      if(PtrToGPIOPort>GPIOC_BASE_ADDRESS)
